Arreglo en la estructura del Codigo
